server/requirements.txt
APScheduler==3.10.4
httpx==0.27.0
openai==1.23.2
Pillow==9.2.0
tornado==6.2
